Remove redundant stuff from gitignore
We don't use Node right now. If/when we do, we can add it.

We also DONT want vscode stuff in the gitignore. That's stuff that belongs in a global gitignore file.
Even the official GitHub gitignore repo says that its recommended to
put the vscode.gitignore in the global

https://github.com/github/gitignore/blob/main/Global/VisualStudioCode.gitignore
https://github.com/github/gitignore#folder-structure